+++ twitter4j-v2-support/src/test/kotlin/twitter4j/ListsTest.kt
@@ -209,6 +209,19 @@ class ListsTest {
     }
 
     @Test
+    fun getListFollowers_simple() {
+
+        // Official Twitter Accounts
+        val listId = 84839422L
+        val res = twitter.getListFollowers(listId)
+        println(res)
+        val json = JSONObject(TwitterObjectFactory.getRawJSON(res))
+        println(json.toString(3))
+
+        assertThat(res.users.size).isGreaterThanOrEqualTo(1)
+    }
+
+    @Test
     fun create_add_member_then_delete() {
 
         println("createList")